rename updates and tests
