+++ src/image_service.py
@@ -1,7 +1,4 @@
-import requests
 import logging
-from .settings import NASA_API_KEY
-
 
 class ImageService:
     def __init__(self, api_client):
@@ -16,13 +13,4 @@ class ImageService:
             "date": response_json["date"],
             "url": response_json["url"],
         }
-
-
-class NasaAPIClient:
-    base_url = "https://api.nasa.gov/planetary/apod"
-    api_key = NASA_API_KEY
-
-    def get_data(self):
-        logging.info("Contacting NASA API")
-        response = requests.get(self.base_url, params={"api_key": self.api_key})
-        return response.json()
+
